+++ bld/f77/wfc/h/scb.h
@@ -2,6 +2,7 @@
 *
 *                            Open Watcom Project
 *
+* Copyright (c) 2002-2021 The Open Watcom Contributors. All Rights Reserved.
 *    Portions Copyright (c) 1983-2002 Sybase, Inc. All Rights Reserved.
 *
 *  ========================================================================
@@ -37,6 +38,6 @@ typedef struct string {         //  run-time string control block
 
 typedef struct cstring {        //  compile-time string control block
     char        *strptr;
-    uint        len;
+    size_t      len;
     char        data;           // used for CHAR( constant )
 } cstring;
